Remove unnecessary install_requires on setuptools (#41)
Remove the explicit runtime dependency on setuptools.  None
of the packages installed by setuptools seem to be used by the package.
While arguably the pytest entry point might require setuptools but then
the dependency belongs (and is present) in pytest, especially that
a future version may use a different entry point loader.