Move third-party library imports to specific tests
Move third-party (`pylibmc` and `MySQLdb`) imports to specific tests,
rather than importing them in global scope.  This makes it easier
to test a subset of the package without having to install all
the dependencies, which can be particularly problematic when wheels
are not provided (e.g. on PyPy3.11 or CPython 3.14).